requirements.txt
numpy
pandas
matplotlib
scikit-learn
keras
tensorflow
streamlit
pillow